Add functions to update the store elsewhere.
I ran into a problem, I can't import any functions in the store
that use the store. The reason is that when those functions are invoked,
the store wouldn't be initialized yet. Inside the two states or values
of store used to update some states using functions that had to be
imported and those functions use the store.

I came up with an idea, I set these values (updateCity and updateScale)
to null. I had functions elsewhere that updated them to appropriate
functions using the store method, updateState. The functions were
invoked in index.js when the store was already initialized.